+++ NEWS.md
@@ -1,3 +1,6 @@
+version 0.3.4
+- updated documentation
+
 version 0.3.2
 - Added a new function cbs_get_data_from_link to allow for retrieving data using a link created with the opendata portal. Thanks to Albert Pieters for the suggestion
 - Improved the base_url default value.